Olhaholiak/favorite_workshops_endpoint (#262)
* add endpoint for user to get favorite workshops

* add pagination

+++ OutOfSchool/OutOfSchool.ElasticsearchData/Models/SearchResultES.cs
@@ -4,7 +4,7 @@ namespace OutOfSchool.ElasticsearchData.Models
 {
     public class SearchResultES<TEntity>
     {
-        public long TotalAmount { get; set; }
+        public int TotalAmount { get; set; }
 
         public IReadOnlyCollection<TEntity> Entities { get; set; }
     }
