+++ OutOfSchool/OutOfSchool.WebApi/Models/SearchResult.cs
@@ -4,7 +4,7 @@ namespace OutOfSchool.WebApi.Models
 {
     public class SearchResult<TEntity>
     {
-        public long TotalAmount { get; set; }
+        public int TotalAmount { get; set; }
 
         public IReadOnlyCollection<TEntity> Entities { get; set; }
     }
